Implement BaseTest class and refactor Before and After Methods in Tests classes
