+++ src/test/java/EpicTest.java
@@ -1,81 +1,19 @@
 import api.ApiManager;
 import api.ApiMethod;
-import api.ApiRequest;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import entities.Epic;
-import entities.Project;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class EpicTest {
-    ApiRequest apiRequest = new ApiRequest();
-    Epic epic = new Epic();
-    Project project = new Project();
-    ConfigFile configFile = new ConfigFile();
+public class EpicTest extends BaseTest {
     int statusNotFound = 404;
     int statusNoContent = 204;
     int statusOk = 200;
     int statusBadRequest = 400;
 
-    @BeforeMethod(onlyForGroups = "createdProject")
-    public void createProjectToTestEpic() throws JsonProcessingException {
-        Project projectTemp = new Project();
-        projectTemp.setName("Project to test in epic");
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("projects");
-        apiRequest.setMethod(ApiMethod.valueOf("POST"));
-        apiRequest.setBody(new ObjectMapper().writeValueAsString(projectTemp));
-        project = ApiManager.executeWithBody(apiRequest).getBody(Project.class);
-    }
-
-    @BeforeMethod(onlyForGroups = "epic")
-    public void setGeneralConfig() throws JsonProcessingException {
-        apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-    }
-
-    @BeforeMethod(onlyForGroups = "createdEpic")
-    public void setCreatedEpicConfig() throws JsonProcessingException {
-        Epic epicTemp = new Epic();
-        epicTemp.setName("Epic to test");
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("projects/{projectId}/epics");
-        apiRequest.setMethod(ApiMethod.valueOf("POST"));
-        apiRequest.addPathParam("projectId", String.valueOf(project.getId()));
-        apiRequest.setBody(new ObjectMapper().writeValueAsString(epicTemp));
-        epic = ApiManager.executeWithBody(apiRequest).getBody(Epic.class);
-    }
-
-    @AfterMethod(onlyForGroups = "deleteCreatedEpic")
-    public void deleteCreatedEpicConfig() {
-        apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
-        apiRequest.setMethod(ApiMethod.DELETE);
-        apiRequest.addPathParam("epicId", String.valueOf(epic.getId()));
-        apiRequest.addPathParam("projectId", String.valueOf(project.getId()));
-        ApiManager.execute(apiRequest);
-    }
-
-    @AfterMethod(onlyForGroups = "deleteProjectOfEpic")
-    public void deleteCreatedProjectToTestEpic() {
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("/projects/{projectId}");
-        apiRequest.setMethod(ApiMethod.DELETE);
-        apiRequest.addPathParam("projectId", String.valueOf(project.getId()));
-        ApiManager.execute(apiRequest);
-    }
-
-    @Test(groups = {"createdProject","createdEpic","epic","deleteCreatedEpic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","createdEpic","setUp","deleteCreatedEpic","deleteCreatedProject"})
     public void itShouldGetAnEpic() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -85,7 +23,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAnEpicWithInvalidId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -95,7 +33,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAnEpicWithInvalidProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -105,7 +43,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAnEpicWithInvalidIdAndProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -115,7 +53,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldGetAllEpicsOfAProject() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/projects/{projectId}/epics");
@@ -124,7 +62,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAllEpicsOfAProjectWithInvalidId() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/projects/{projectId}/epics");
@@ -133,7 +71,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAllEpicsOfAProjectWithoutId() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/projects/{projectId}/epics");
@@ -142,7 +80,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","createdEpic","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","createdEpic","setUp","deleteCreatedProject"})
     public void itShouldDeleteAnEpic() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -152,7 +90,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNoContent);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteAnEpicWithInvalidId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -162,7 +100,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusBadRequest);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteAnEpicWithoutId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -172,7 +110,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteAnEpicWithoutProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -182,7 +120,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteAnEpicWithoutIdNorProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -192,7 +130,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteAnEpicWithInvalidProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -202,7 +140,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteAnEpicWithInvalidIdAndInvalidProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/epics/{epicId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -212,7 +150,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteCreatedEpic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedEpic","deleteCreatedProject"})
     public void itShouldCreateAnEpic() throws JsonProcessingException {
         Epic epicTemp = new Epic();
         epicTemp.setName("Epic created");
@@ -226,7 +164,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotCreateAnEpicWithEmptyName() throws JsonProcessingException {
         Epic epicTemp = new Epic();
         epicTemp.setName("");
@@ -239,7 +177,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusBadRequest);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotCreateAnEpicWithoutProjectId() throws JsonProcessingException {
         Epic epicTemp = new Epic();
         epicTemp.setName("Epic created");
@@ -252,7 +190,7 @@ public class EpicTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","epic","deleteProjectOfEpic"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotCreateAnEpicWithInvalidProjectId() throws JsonProcessingException {
         Epic epicTemp = new Epic();
         epicTemp.setName("Epic created");
