+++ src/test/java/LabelTest.java
@@ -1,81 +1,19 @@
 import api.ApiManager;
 import api.ApiMethod;
-import api.ApiRequest;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import entities.Label;
-import entities.Project;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class LabelTest {
-    ApiRequest apiRequest = new ApiRequest();
-    Label label = new Label();
-    Project project = new Project();
-    ConfigFile configFile = new ConfigFile();
+public class LabelTest extends BaseTest {
     int statusNotFound = 404;
     int statusNoContent = 204;
     int statusOk = 200;
     int statusBadRequest = 400;
 
-    @BeforeMethod(onlyForGroups = "createdProject")
-    public void createProjectToTestLabel() throws JsonProcessingException {
-        Project projectTemp = new Project();
-        projectTemp.setName("Project to test in label");
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("projects");
-        apiRequest.setMethod(ApiMethod.valueOf("POST"));
-        apiRequest.setBody(new ObjectMapper().writeValueAsString(projectTemp));
-        project = ApiManager.executeWithBody(apiRequest).getBody(Project.class);
-    }
-
-    @BeforeMethod(onlyForGroups = "label")
-    public void setGeneralConfig() throws JsonProcessingException {
-        apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-    }
-
-    @BeforeMethod(onlyForGroups = "createdLabel")
-    public void setCreatedLabelConfig() throws JsonProcessingException {
-        Label labelTemp = new Label();
-        labelTemp.setName("Label to test");
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("projects/{projectId}/labels");
-        apiRequest.setMethod(ApiMethod.valueOf("POST"));
-        apiRequest.addPathParam("projectId", String.valueOf(project.getId()));
-        apiRequest.setBody(new ObjectMapper().writeValueAsString(labelTemp));
-        label = ApiManager.executeWithBody(apiRequest).getBody(Label.class);
-    }
-
-    @AfterMethod(onlyForGroups = "deleteCreatedLabel")
-    public void deleteCreatedLabelConfig() {
-        apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
-        apiRequest.setMethod(ApiMethod.DELETE);
-        apiRequest.addPathParam("labelId", String.valueOf(label.getId()));
-        apiRequest.addPathParam("projectId", String.valueOf(project.getId()));
-        ApiManager.execute(apiRequest);
-    }
-
-    @AfterMethod(onlyForGroups = "deleteProjectOfLabel")
-    public void deleteCreatedProjectToTestLabel() {
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("/projects/{projectId}");
-        apiRequest.setMethod(ApiMethod.DELETE);
-        apiRequest.addPathParam("projectId", String.valueOf(project.getId()));
-        ApiManager.execute(apiRequest);
-    }
-
-    @Test(groups = {"createdProject","createdLabel","label","deleteCreatedLabel","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","createdLabel","setUp","deleteCreatedLabel","deleteCreatedProject"})
     public void itShouldGetALabel() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -85,7 +23,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetALabelWithInvalidId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -95,7 +33,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetALabelWithInvalidProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -105,7 +43,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetALabelWithInvalidIdAndProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -115,7 +53,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldGetAllLabelsOfAProject() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/projects/{projectId}/labels");
@@ -124,7 +62,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAllLabelsOfAProjectWithInvalidId() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/projects/{projectId}/labels");
@@ -133,7 +71,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotGetAllLabelsOfAProjectWithoutId() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/projects/{projectId}/labels");
@@ -142,7 +80,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","createdLabel","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","createdLabel","setUp","deleteCreatedProject"})
     public void itShouldDeleteALabel() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -152,7 +90,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNoContent);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteALabelWithInvalidId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -162,7 +100,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusBadRequest);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteALabelWithoutId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -172,7 +110,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteALabelWithoutProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -182,7 +120,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteALabelWithoutIdNorProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -192,7 +130,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteALabelWithInvalidProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -202,7 +140,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotDeleteALabelWithInvalidIdAndInvalidProjectId() {
         apiRequest.setEndpoint("/projects/{projectId}/labels/{labelId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -212,7 +150,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteCreatedLabel","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedLabel","deleteCreatedProject"})
     public void itShouldCreateALabel() throws JsonProcessingException {
         Label labelTemp = new Label();
         labelTemp.setName("Label created");
@@ -226,7 +164,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotCreateALabelWithEmptyName() throws JsonProcessingException {
         Label labelTemp = new Label();
         labelTemp.setName("");
@@ -239,7 +177,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusBadRequest);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotCreateALabelWithoutProjectId() throws JsonProcessingException {
         Label labelTemp = new Label();
         labelTemp.setName("Label created");
@@ -252,7 +190,7 @@ public class LabelTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"createdProject","label","deleteProjectOfLabel"})
+    @Test(groups = {"createdProject","setUp","deleteCreatedProject"})
     public void itShouldNotCreateALabelWithInvalidProjectId() throws JsonProcessingException {
         Label labelTemp = new Label();
         labelTemp.setName("Label created");
