+++ src/test/java/WorkspaceTest.java
@@ -1,53 +1,19 @@
 import api.ApiManager;
 import api.ApiMethod;
-import api.ApiRequest;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import entities.Workspace;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public class WorkspaceTest {
-    ApiRequest apiRequest = new ApiRequest();
-    Workspace workspace = new Workspace();
-    ConfigFile configFile = new ConfigFile();
+public class WorkspaceTest extends BaseTest {
     int statusNotFound = 404;
     int statusNoContent = 204;
     int statusOk = 200;
     int statusBadRequest = 400;
 
-    @BeforeMethod(onlyForGroups = "workspace")
-    public void setGeneralConfig() {
-        apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-    }
-
-    @BeforeMethod(onlyForGroups = "createdWorkspace")
-    public void setCreatedWorkspaceConfig() throws JsonProcessingException {
-        Workspace workspaceTemp = new Workspace();
-        workspaceTemp.setName("Workspace to test");
-        ApiRequest apiRequest = new ApiRequest();
-        apiRequest.addHeader("X-TrackerToken", configFile.getConfig().getProperty("PIVOTAL_TOKEN"));
-        apiRequest.setBaseUri(configFile.getConfig().getProperty("PIVOTAL_BASE_URI"));
-        apiRequest.setEndpoint("/my/workspaces");
-        apiRequest.setMethod(ApiMethod.valueOf("POST"));
-        apiRequest.setBody(new ObjectMapper().writeValueAsString(workspaceTemp));
-        workspace = ApiManager.executeWithBody(apiRequest).getBody(Workspace.class);
-    }
-
-    @AfterMethod(onlyForGroups = "deleteCreatedWorkspace")
-    public void deleteCreatedWorkspaceConfig() {
-        apiRequest.setEndpoint("/my/workspaces/{workspaceId}");
-        apiRequest.setMethod(ApiMethod.DELETE);
-        apiRequest.addPathParam("workspaceId", String.valueOf(workspace.getId()));
-        ApiManager.execute(apiRequest);
-    }
-
-    @Test(groups = "workspace")
+    @Test(groups = "setUp")
     public void itShouldGetAllWorkspaces() {
         apiRequest.setMethod(ApiMethod.GET);
         apiRequest.setEndpoint("/my/workspaces");
@@ -55,7 +21,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = {"createdWorkspace","workspace","deleteCreatedWorkspace"})
+    @Test(groups = {"createdWorkspace","setUp","deleteCreatedWorkspace"})
     public void itShouldGetAWorkspace() {
         apiRequest.setEndpoint("/my/workspaces/{workspaceId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -64,7 +30,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = "workspace")
+    @Test(groups = "setUp")
     public void itShouldNotGetAWorkspaceWithInvalidId() {
         apiRequest.setEndpoint("/my/workspaces/{workspaceId}");
         apiRequest.setMethod(ApiMethod.GET);
@@ -73,7 +39,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = {"workspace","deleteCreatedWorkspace"})
+    @Test(groups = {"setUp","deleteCreatedWorkspace"})
     public void itShouldCreateAWorkspace() throws JsonProcessingException {
         Workspace workspaceTemp = new Workspace();
         workspaceTemp.setName("Workspace created");
@@ -86,7 +52,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusOk);
     }
 
-    @Test(groups = "workspace")
+    @Test(groups = "setUp")
     public void itShouldNotCreateAWorkspaceWithEmptyName() throws JsonProcessingException {
         Workspace workspaceTemp = new Workspace();
         workspaceTemp.setName("");
@@ -98,7 +64,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusBadRequest);
     }
 
-    @Test(groups = {"createdWorkspace","workspace"})
+    @Test(groups = {"createdWorkspace","setUp"})
     public void itShouldDeleteWorkspace() {
         apiRequest.setEndpoint("/my/workspaces/{workspaceId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -107,7 +73,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNoContent);
     }
 
-    @Test(groups = "workspace")
+    @Test(groups = "setUp")
     public void itShouldNotDeleteAWorkspaceWithInvalidId() {
         apiRequest.setEndpoint("/my/workspaces/{workspaceId}");
         apiRequest.setMethod(ApiMethod.DELETE);
@@ -116,7 +82,7 @@ public class WorkspaceTest {
         Assert.assertEquals(apiResponse.getStatusCode(), statusNotFound);
     }
 
-    @Test(groups = "workspace")
+    @Test(groups = "setUp")
     public void itShouldNotDeleteAWorkspaceWithoutId() {
         apiRequest.setEndpoint("/my/workspaces/{workspaceId}");
         apiRequest.setMethod(ApiMethod.DELETE);
